[refact] move actor/critic/hybrid_engine/reward_model/rollout/workers out of verl/trainer/ppo directory to verl/trainer directory for reusage
